data-pipeline/requirements.txt
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.1.0
numpy==2.2.1
scikit-learn==1.6.1
qdrant-client==1.12.1
neo4j==5.27.0
httpx==0.28.1
sentence-transformers==3.3.1
python-dotenv==1.0.1
tqdm==4.67.1
